refactor: move ApiKeyScope into shorebird_code_push_protocol
The scope vocabulary is not CLI-specific. The auth service mints a key
against one of these presets and code_push_server is what enforces the
resulting permissions on every request the key makes, so the term is shared
between server and client — which is what the protocol package is for.

It lands in `lib/extensions/` rather than `lib/src/` because `lib/src/` is
generated from the CodePush OpenAPI spec and API keys are minted by the auth
service, which that spec does not cover. `AuthProvider` sits there for the
same reason.

Only the wire value moves. `--scope`'s hyphenated spelling is a command-line
concern and stays in the CLI as an extension.
